+++ bitrix/modules/main/lib/userprofilerecord.php
@@ -10,6 +10,24 @@ namespace Bitrix\Main;
 
 use Bitrix\Main\Entity;
 
+/**
+ * @internal
+ * Class UserProfileRecordTable
+ * @package Bitrix\Main
+ *
+ * DO NOT WRITE ANYTHING BELOW THIS
+ *
+ * <<< ORMENTITYANNOTATION
+ * @method static EO_UserProfileRecord_Query query()
+ * @method static EO_UserProfileRecord_Result getByPrimary($primary, array $parameters = array())
+ * @method static EO_UserProfileRecord_Result getById($id)
+ * @method static EO_UserProfileRecord_Result getList(array $parameters = array())
+ * @method static EO_UserProfileRecord_Entity getEntity()
+ * @method static \Bitrix\Main\EO_UserProfileRecord createObject($setDefaultValues = true)
+ * @method static \Bitrix\Main\EO_UserProfileRecord_Collection createCollection()
+ * @method static \Bitrix\Main\EO_UserProfileRecord wakeUpObject($row)
+ * @method static \Bitrix\Main\EO_UserProfileRecord_Collection wakeUpCollection($rows)
+ */
 class UserProfileRecordTable extends Entity\DataManager
 {
 	public static function getTableName()
@@ -39,18 +57,21 @@ class UserProfileRecordTable extends Entity\DataManager
 		);
 	}
 
-	public static function deleteByUser($userId)
+	public static function deleteByHistoryFilter($where)
 	{
-		$userId = intval($userId);
+		if($where == '')
+		{
+			throw new ArgumentException("Deleting by empty filter is not allowed, use truncate (b_user_profile_record).", "where");
+		}
 
 		$entity = static::getEntity();
 		$conn = $entity->getConnection();
 
-		$conn->queryExecute("
+		$conn->query("
 			DELETE FROM b_user_profile_record 
 			WHERE HISTORY_ID IN(
 				SELECT ID FROM b_user_profile_history 
-				WHERE USER_ID = {$userId}
+				{$where} 
 			)"
 		);
